Use node-pre-gyp to (pre-)compile binaries
node-pre-gyp first checks for a precompiled node binding at a
configurable URL and falls back to building it from scratch if
there is no file at the given URL.

There's also AWS S3 support for hosting files, but for now, let's
use GitHub releases for that.

To provide a precompile archive use `npm run-script package` and
upload the generated archive (build/stage/...../netif-*.tar.gz) on
the GitHub releases page. This step can also be automated with
Travis and Appveyor.