Don't depend on specific interface names in tests.
Systems tend to have wildly varying names for their interfaces.
(Windows in particular.) So instead of guessing based on the
operating system, use `os.networkInterfaces()` to get a list of
interfaces.

Also, use a specific nodeunit version, because 0.9 can not be installed
on node 0.8 since it uses the caret (^) notation in it's `package.json`.