Windows compatibility.
Not as hard as it seemed. Caveat: Queries up to 16 interfaces.
Throws an error if the host has more interfaces than fit inside
the array.  Can be fixed by dynamically allocating the
IP_ADAPTER_ADDRESSES array and retrying if `GetAdaptersAddresses`
returns ERROR_BUFFER_OVERFLOW.

Also, as the documentation states, this function "requires a
significant amount of network resources and time", so we should
probably keep the list of addresses around.